Split agent.go into per-subcommand files (agent_*.go)
agent.go had grown to 677 lines with 11 subcommands, the poll loop,
error rendering helpers, and shared types all stacked together. Split
into one file per subcommand using the flat-prefixed pattern (Pattern B
from the review):

  agent.go              — parent cmd + init() + shared types + helpers
  agent_create.go       — agentCreateCmd + pollUntilTerminal
  agent_get.go          — agentGetCmd
  agent_list.go         — agentListCmd
  agent_delete.go       — agentDeleteCmd
  agent_connect.go      — connect + disconnect + channels
  agent_packages.go     — install + uninstall + packages + renderAgentError
  agent_events.go       — events command + row types + truncate/valueOr
  agent_errors.go       — LastError + ExitError + codeCatalog + renderers
                          (was errors.go — misleading, looked like it should
                           register an `oc errors` command)

Why Pattern B (flat, prefixed) instead of a subpackage:
- Matches existing convention — sandbox.go, exec.go, secret.go are all
  single files in this package.
- Private helpers (sessionsClient, formatList, formatAge, printer,
  jsonOutput) stay private to the package; no export ceremony.
- When sandbox.go or exec.go eventually outgrow themselves, they can
  follow the same prefix pattern without introducing a precedent break.

Scope: agent commands only. sandbox.go / exec.go / secret.go /
checkpoint.go untouched.

Also removed now-unused instanceResponse / instanceListResponse types —
the second fetch to /instances was dropped when GET /v1/agents/:id
started returning instance status + last_error inline.
